refactor the component repo structure
