+++ src/test/java/com/vaadin/ui/radiobutton/tests/RadioButtonGroupIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.radiobutton;
+package com.vaadin.ui.radiobutton.tests;
 
 import java.util.List;
